Create runtime package and entrypoint script
Prime CMS needs the compiled node modules to be required by a runtime
package.
The entrypoint script is generating a new SESSION_SECRET on
every launch if the user does not provide a value. It is currently
also trying to fix issue 117
(https://github.com/birkir/prime/issues/117) with a quickfix. It does
not work in all cases!